Added new API for validating webapps without files (bug 772857)
You can now validate add-ons with the validate.validate_app() function
instead of validate.validate. It's faster because it cuts through a
lot of cruft, but it also enables multiple-datacenter validation
becaues it doesn't rely on local files. The string version of the
JSON is passed to the API instead of a path and the result is
returned as a JSON string.t push