Remove package.json and manifest.json parsing: it was not used
Since it was never loaded by validator/loader.py the package.json
and manifest.json parsing code was never used... except in tests.
This could cause a valid extension to be considered invalid only
in tests.

We don't need to parse package.json anyway, jpm generates an
install.rdf and that's what we test. We do need to parse
manifest.json but we no longer use amo-validator for webextensions
and rely on addons-linter instead.